cognicare-connect-main/cognicare-connect-main/ml-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
numpy>=1.26.0
Pillow>=10.4.0
tensorflow>=2.18.0
huggingface-hub>=0.24.0
